add setup.py: configuration for tools such as pip
The standard way to install Python packages is via the ``pip install`` command.
This command ultimately ends up running the ``setup.py`` script in order to
determine:

1. Which additional PyPI packages we required
2. Which packages we provide
3. Metadata on out package such as name, description and version

Add a very basic stub setup.py which takes advantage of the auto-discovery
capabilities of Python if one lays things out in a standard manner.