add simple pytest test example
Tests are simple python files. Each function named ``test_...`` is called in
turn. Any which throw an exception are classed as failing. Pytest includes a
great deal of utility functions which can help with writing tests. See the
documentation at: http://pytest.org/.

Add a simple test suite which contains a single test which always succeeds. Not
the most rigorous of test suites :).