travis: add rules to build documentation
Add a new tox testenv, "doc", which will install sphinx and build the
documentation. Add configuration to .travis.yml to test the doc test env as well
as the previous ones.

Testing the examples and the test suite as one is a little wasteful since they
can easily be parallelised. Split tests and examples out into separate tox envs.