tox: enable a linting test pass
Add a linting pass which uses pylint to check code style issues in cusfsim. For
the moment do not include examples or tests since they may have specific reasons
to be non-idiomatic.

Some lint checks have been disabled via pylintrc. The idea of this is to try and
reduce lint noise for checks which are subjective or better done in code review.
This list is likely non-exhaustive and may need updating.