Documents and tests the new kinematics functionality. Fixes #40 and fixes #41 in theory.
The new kinematics functionality is tested to make sure that it is correctly changing the mass of the rocket and agrees with a 1d analytical solution for cosntant acceleration.

Documentation has been added for a kinematics_example.py which repeats what the original kinematics.py did but in the new framework.